# Change: Check trash target references before accessing them