## What

In the trash case of GET_TARGETS check that dependent resources exist before trying to access info about them.

Specifically, in `handle_get_targets` check `ssh_credential`, `smb_credential`, etc before passing them to functions like `trash_credential_name`.

## Why

This saves the SQL statements of `trash_credential_name` etc when `ssh_credential` etc are 0. These statements would have always returned NULL.

Note that this only affects the trash case, because the regular case was already doing the check.

The speedup looks mild if any, but we might as well skip the extra SQL. With 1600 trash targets:
```
time o m m '<get_targets filter="rows=-1 sort=name" trash="1"/>' > /tmp/targ
before: 9s,8s,7s
after: 7s,6s,6s
```